Updating to conform with latest buster-configuration API.
onConfig simply resolves, the 'value' of the config is the list of groups so we iterate by hand.

The config API also provides a list of the absolute path to all load entries so we don't have to compute that here.

Note, the browser runner has not been updated at all, so it currently breaks miserably.