Disabling the ... animation in the header, it causes chrome to lock up.
Perhaps this applies to other browsers as well, but it seems that as soon as the capture web page tab is not in the foreground, faye is no longer able to do networking properly and chokes. Perhaps this is because faye also uses timeout and chrome only allows one timeout at a time, prioritizing the animation timeout. Haven't really investigated.